Some renameing/refactoring and adding check for SDK version
