+++ build.sbt
@@ -7,6 +7,15 @@ scalaVersion := "2.11.12"
 organization := "io.kf.etl"
 version := "1.0.0"
 
+javacOptions ++= Seq("-source", "1.8", "-target", "1.8", "-Xlint")
+
+initialize := {
+  val _ = initialize.value
+  val javaVersion = sys.props("java.specification.version")
+  if (javaVersion != "1.8")
+    sys.error("Java 1.8 is required for this project. Found " + javaVersion + " instead")
+}
+
 resolvers ++= Seq(
   clojars,
   maven_local,
