+++ src/main/scala/io/kf/etl/processors/participantcentric/transform/ParticipantCentricTransformer.scala
@@ -1,140 +1,122 @@
 package io.kf.etl.processors.participantcentric.transform
 
-import io.kf.etl.models.es.{ParticipantCentric_ES, Participant_ES}
-import io.kf.etl.models.internal.{BiospecimenId_GenomicFileId, SequencingExperimentsES_GenomicFileId, _}
+import io.kf.etl.models.dataservice.{EBiospecimen, EBiospecimenGenomicFile}
+import io.kf.etl.models.es.{BiospecimenCombined_ES, GenomicFile_ES, ParticipantCombined_ES, Participant_ES}
+import io.kf.etl.models.internal.{SequencingExperimentsES_GenomicFileId, _}
 import io.kf.etl.processors.common.ProcessorCommonDefinitions.EntityDataSet
 import io.kf.etl.processors.common.converter.EntityConverter
 import org.apache.spark.sql.{Dataset, SparkSession}
 
 object ParticipantCentricTransformer{
 
-  def apply(entityDataset:EntityDataSet, participants: Dataset[Participant_ES])(implicit spark:SparkSession): Dataset[ParticipantCentric_ES] = {
+  def apply(entityDataset:EntityDataSet, participants: Dataset[Participant_ES])(implicit spark:SparkSession): Dataset[ParticipantCombined_ES] = {
     import spark.implicits._
 
-    val fileId_experiments = entityDataset.sequencingExperiments
-      .joinWith(
-        entityDataset.sequencingExperimentGenomicFiles,
-        entityDataset.sequencingExperiments.col("kfId") === entityDataset.sequencingExperimentGenomicFiles.col("sequencingExperiment"),
-        "left_outer"
-      )
-      .map(tuple => {
-        SequencingExperimentES_GenomicFileId(
-          sequencingExperiment = EntityConverter.ESequencingExperimentToSequencingExperimentES(tuple._1),
-          genomicFileId = tuple._2.genomicFile
+
+    val fileId_experiments: Dataset[SequencingExperimentsES_GenomicFileId] =
+      entityDataset.sequencingExperiments
+        .joinWith(
+          entityDataset.sequencingExperimentGenomicFiles,
+          entityDataset.sequencingExperiments.col("kfId") === entityDataset.sequencingExperimentGenomicFiles.col("sequencingExperiment"),
+          "left_outer"
         )
-      })
-      .groupByKey(_.genomicFileId)
-      .mapGroups((fileId, iterator) => {
-        fileId match {
-          case Some(id) =>
+        .map(tuple => {
+          SequencingExperimentES_GenomicFileId(
+            sequencingExperiment = EntityConverter.ESequencingExperimentToSequencingExperimentES(tuple._1),
+            genomicFileId = tuple._2.genomicFile
+          )
+        })
+        .groupByKey(_.genomicFileId)
+        .mapGroups((fileId, iterator) => {
+          fileId match {
+            case Some(id) =>
 
-            val experiments = iterator.map(_.sequencingExperiment).toSeq
-            SequencingExperimentsES_GenomicFileId(
-              sequencingExperiments = experiments,
-              genomicFileId = id
-            )
-          case None => null
-        }
-      })
-      .filter(_!=null)
+              val experiments = iterator.map(_.sequencingExperiment).toSeq
+              SequencingExperimentsES_GenomicFileId(
+                genomicFileId = id,
+                sequencingExperiments = experiments
+              )
+            case None => null
+          }
+        })
+        .filter(_ != null)
 
-    val files =
+    val files: Dataset[GenomicFile_ES] =
       entityDataset.genomicFiles.joinWith(
         fileId_experiments,
-        entityDataset.genomicFiles.col("kfId") === fileId_experiments.col("genomicFile"),
+        entityDataset.genomicFiles.col("kfId") === fileId_experiments.col("genomicFileId"),
         "left_outer"
       ).map(tuple => {
         Option(tuple._2) match {
           case Some(_) =>
-            EntityConverter.EGenomicFileToGenomicFileES(tuple._1, tuple._2.sequencingExperiments)
+            EntityConverter.EGenomicFileToGenomicFileES(
+              tuple._1,
+              tuple._2.sequencingExperiments)
           case None => EntityConverter.EGenomicFileToGenomicFileES(tuple._1, Seq.empty)
         }
       })
 
-    val bio_gf: Dataset[BiospecimenId_GenomicFileId] =
-      entityDataset.biospecimens
-        .joinWith(
-          entityDataset.biospecimenGenomicFiles,
-          entityDataset.biospecimens.col("kfId") === entityDataset.biospecimenGenomicFiles.col("biospecimenId"),
-          "left_outer"
-        )
-        .map(tuple => {
-          BiospecimenId_GenomicFileId(
-            bioId = tuple._1.kfId,
-            gfId = {
-              Option(tuple._2) match {
-                case Some(_) => tuple._2.genomicFileId
-                case None => null
-              }
-            }
+    val bio_gf: Dataset[(EBiospecimen, Seq[GenomicFile_ES])] = {
+      val bioSpec_GFs: Dataset[(EBiospecimen, GenomicFile_ES)] =
+        entityDataset.biospecimens
+          .joinWith(
+            entityDataset.biospecimenGenomicFiles,
+            entityDataset.biospecimens.col("kfId") === entityDataset.biospecimenGenomicFiles.col("biospecimenId"),
+            "left_outer"
           )
-        })
+          .as[(EBiospecimen, EBiospecimenGenomicFile)]
+          .toDF("eBiospecimen", "eBiospecimenGenomicFile")
+          .joinWith(
+            files,
+            $"eBiospecimenGenomicFile.genomicFileId" === files("kf_id")
+          )
+          .as[((EBiospecimen, EBiospecimenGenomicFile), GenomicFile_ES)]
+          .map { case ((biospecimen, _), file) => (biospecimen, file) }
 
-
-    val bioId_File =
-      bio_gf.joinWith(
-        files,
-        bio_gf.col("gfId") === files.col("kf_id")
-      ).map(tuple => {
-        BiospecimenId_GenomicFileES(
-          bioId = tuple._1.bioId.get,
-          file = tuple._2
-        )
-      })
-
-
-    val participants_bioId =
-      participants.joinWith(
-        entityDataset.biospecimens,
-        participants.col("kf_id") === entityDataset.biospecimens.col("participantId"),
-        "left_outer"
-      ).flatMap(tuple => {
-
-        Option(tuple._2) match {
-          case Some(_) =>
-            Seq(ParticipantES_BiospecimenId(bioId = tuple._2.kfId, participant = tuple._1))
-          case None =>
-            Seq(ParticipantES_BiospecimenId(participant = tuple._1, bioId = None))
+      bioSpec_GFs.groupByKey { case (biospecimen, _) => biospecimen.kfId }
+        .mapGroups { case (_, groupsIterator) =>
+          val groups = groupsIterator.toSeq
+          val biospecimen: EBiospecimen = groups.head._1
+          val files: Seq[GenomicFile_ES] = groups.map(_._2)
+          (biospecimen, files)
         }
+    }
 
-      })
-
-
-    participants_bioId.joinWith(
-      bioId_File,
-      participants_bioId.col("bioId") === bioId_File.col("bioId"),
+    participants.joinWith(
+      bio_gf,
+      participants.col("kf_id") === bio_gf.col("_1.participantId"),
       "left_outer"
-    ).groupByKey(tuple => {
-      tuple._1.participant.kf_id.get
-    }).mapGroups((_, iterator) => {
-      val seq = iterator.toSeq
+    ).groupByKey { case (participant, _) => participant.kf_id.get }
+      .mapGroups { case (_, groupsIterator) =>
+        val groups = groupsIterator.toSeq
+        val participant = groups.head._1
+        val bioFiles: Seq[BiospecimenCombined_ES] = groups.collect {
+          case (_, (biospecimen, gfiles)) if biospecimen != null => EntityConverter.EBiospecimenToBiospecimenCombinedES(biospecimen, gfiles)
+        }
+        val gfiles: Seq[GenomicFile_ES] = bioFiles.flatMap(_.genomic_files)
 
-      val participant = seq.head._1.participant
 
-      ParticipantCentric_ES(
-        affected_status = participant.affected_status,
-        alias_group = participant.alias_group,
-        biospecimens = Nil, //participant.biospecimens,
-        diagnoses = participant.diagnoses,
-        ethnicity = participant.ethnicity,
-        external_id = participant.external_id,
-        family = participant.family,
-        family_id = participant.family_id,
-        files = seq.collect{
-          case tuple if tuple._2 != null => tuple._2.file
-        },
-        gender = participant.gender,
-        is_proband = participant.is_proband,
-        kf_id = participant.kf_id,
-        outcome = participant.outcome,
-        phenotype = participant.phenotype,
-        race = participant.race,
-        study = participant.study,
-        available_data_types = participant.available_data_types
-      )
-    })
+        ParticipantCombined_ES(
+          affected_status = participant.affected_status,
+          alias_group = participant.alias_group,
+          available_data_types = participant.available_data_types,
+          biospecimens = bioFiles,
+          diagnoses = participant.diagnoses,
+          diagnosis_category = participant.diagnosis_category,
+          ethnicity = participant.ethnicity,
+          external_id = participant.external_id,
+          family = participant.family,
+          family_id = participant.family_id,
+          files = gfiles,
+          gender = participant.gender,
+          is_proband = participant.is_proband,
+          kf_id = participant.kf_id,
+          outcome = participant.outcome,
+          phenotype = participant.phenotype,
+          race = participant.race,
+          study = participant.study
+        )
+      }
 
   }
-
-
 }
